LayoutTests/imported/w3c/web-platform-tests/intersection-observer/clip-path-expected.txt


FAIL clip-path assert_equals: entries.length expected 2 but got 1
